scripts/requirements.txt
google-api-python-client>=2.100
python-dotenv>=1.0
requests>=2.31
